changed uri and added more tests
