+++ api/db/services/document_service.py
@@ -61,23 +61,25 @@ class DocumentService(CommonService):
 
     @classmethod
     @DB.connection_context()
-    def get_by_kb_id_by_offset(cls, kb_id, offset, count, orderby, desc, keywords):
+    def get_documents_by_dataset_id(cls, dataset_id, offset, count, order_by, descend, keywords):
         if keywords:
             docs = cls.model.select().where(
-                (cls.model.kb_id == kb_id),
+                (cls.model.kb_id == dataset_id),
                 (fn.LOWER(cls.model.name).contains(keywords.lower()))
             )
         else:
-            docs = cls.model.select().where(cls.model.kb_id == kb_id)
+            docs = cls.model.select().where(cls.model.kb_id == dataset_id)
 
         total = docs.count()
-        if desc:
-            docs = docs.order_by(cls.model.getter_by(orderby).desc())
-        else:
-            docs = docs.order_by(cls.model.getter_by(orderby).asc())
+
+        if descend == 'True':
+            docs = docs.order_by(cls.model.getter_by(order_by).desc())
+        if descend == 'False':
+            docs = docs.order_by(cls.model.getter_by(order_by).asc())
 
         docs = list(docs.dicts())
         docs_length = len(docs)
+
         if offset < 0 or offset > docs_length:
             raise IndexError("Offset is out of the valid range.")
